// src/dota2/types.ts
export type Match = {
    players: Player[];
    radiant_win: boolean;
    duration: number;
    pre_game_duration: number;
    start_time: number;
    match_id: number;
    match_seq_num: number;
    tower_status_radiant: number;
    tower_status_dire: number;
    barracks_status_radiant: number;
    barracks_status_dire: number;
    cluster: number;
    first_blood_time: number;
    lobby_type: number;
    human_players: number;
    leagueid: number;
    game_mode: number;
    flags: number;
    engine: number;
    radiant_score: number;
    dire_score: number;
    picks_bans: PickBan[];
};

export type Player = {
    account_id: number;
    player_slot: number;
    team_number: number;
    team_slot: number;
    hero_id: number;
    hero_variant: number;
    item_0: number;
    item_1: number;
    item_2: number;
    item_3: number;
    item_4: number;
    item_5: number;
    backpack_0: number;
    backpack_1: number;
    backpack_2: number;
    item_neutral: number;
    kills: number;
    deaths: number;
    assists: number;
    leaver_status: number;
    last_hits: number;
    denies: number;
    gold_per_min: number;
    xp_per_min: number;
    level: number;
    net_worth: number;
    aghanims_scepter: number;
    aghanims_shard: number;
    moonshard: number;
    hero_damage: number;
    tower_damage: number;
    hero_healing: number;
    gold: number;
    gold_spent: number;
    ability_upgrades_arr: number[];
    personaname: string | null;
    name: string | null;
    last_login: string | null;
    radiant_win: boolean;
    start_time: number;
    duration: number;
    cluster: number;
    lobby_type: number;
    game_mode: number;
    is_contributor: boolean;
    patch: number;
    region: number;
    isRadiant: boolean;
    win: number;
    lose: number;
    total_gold: number;
    total_xp: number;
    kills_per_min: number;
    kda: number;
    abandons: number;
    rank_tier: number | null;
    is_subscriber: boolean;
    benchmarks: Benchmarks;
};

export type Benchmarks = {
    gold_per_min: Benchmark;
    xp_per_min: Benchmark;
    kills_per_min: Benchmark;
    last_hits_per_min: Benchmark;
    hero_damage_per_min: Benchmark;
    hero_healing_per_min: Benchmark;
    tower_damage: Benchmark;
};

export type Benchmark = {
    raw: number;
    pct: number;
};

export type PickBan = {
    is_pick: boolean;
    hero_id: number;
    team: number;
    order: number;
};